requirements.txt
fastapi
pydantic
python-multipart
openai
streamlit
numpy
pdfplumber
python-dotenv